Link `ftPr_Init` (1.74 KB)
